% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annual_mortality_rates_2015}
\alias{annual_mortality_rates_2015}
\title{Human mortality database age and sex specific rates for all cause deaths}
\format{
\subsection{\code{annual_mortality_rates_2015}}{

a data.table with 3 rows and 113 columns.
\describe{
\item{birth_cohort}{Birth cohort as a calendar year}
\item{sex}{sex: \code{female}, \code{male}, \code{total}.}
\item{age_0, ..., age 110+}{age-specific death rates}
}
}
}
\source{
\url{https://www.mortality.org}
}
\usage{
annual_mortality_rates_2015
}
\description{
This is the 2015 annual death rates from the 2023 version of the
Human Mortality Database for the USA.
}
\keyword{internal}
