% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper.R
\name{compare_ppp_vectors}
\alias{compare_ppp_vectors}
\title{Check that two ppp vectors Q-Q agree}
\usage{
compare_ppp_vectors(ppp1, ppp2, threshold = 0.15, showQQ = TRUE)
}
\arguments{
\item{ppp1}{(vector, double) the first vector}

\item{ppp2}{(vector, double) the second vector}

\item{threshold}{(double) optional: the cutoff for a large absolute threshold}

\item{showQQ}{(boolean) optional: show the QQ plot if the absolute value of the
Difference vs Average ratio in any decile is bigger than the \code{threshold}}
}
\value{
None
}
\description{
Compare that the deciles of two vectors have absolute difference
over average ratios less than \code{threshold}
}
