% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_t_intensity_exponential.R
\name{draw_intensity_exponential}
\alias{draw_intensity_exponential}
\title{Simulate from a non homogeneous Poisson Point Process (NHPPP) from
(t_min, t_max) with exponential intensity function (inversion method)}
\usage{
draw_intensity_exponential(
  alpha = 1,
  beta = 0,
  range_t = c(0, 10),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{alpha}{(double) the intercept in the exponent}

\item{beta}{(double) the slope in the exponent}

\item{range_t}{(vector, double) min and max of the time interval}

\item{rng_stream}{(\code{rstream}) an \code{rstream} object.}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times (t_); if no events realize,
a vector of length 0
}
\description{
Sample NHPPP times from an exponential intensity function
using the inversion method, optionally using an \code{rstream}
generator or a \code{Kystis} \code{RNGClass} object
}
\examples{
x <- draw_intensity_exponential(alpha = 0, beta = 0.2)

}
