% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_zt_intensity_loglinear.R
\name{draw_zt_intensity_loglinear}
\alias{draw_zt_intensity_loglinear}
\title{Simulate from a zero-truncated non homogeneous Poisson Point Process (zt-NHPPP) from
(t_min, t_max) with a log-linear intensity function (inversion method)}
\usage{
draw_zt_intensity_loglinear(
  alpha = 1,
  beta = 0,
  range_t = c(0, 10),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{alpha}{(double) the intercept in the exponent}

\item{beta}{(double) the slope in the exponent}

\item{range_t}{(vector, double) min and max of the time interval}

\item{rng_stream}{(\code{rstream}) an \code{rstream} object.}

\item{only1}{boolean, 1 event time}
}
\value{
a vector of at least 1 event times
}
\description{
Sample zt-NHPPP times from an log-linear intensity function
using the inversion method, optionally using an \code{rstream}
generator
}
\examples{
x <- draw_zt_intensity_loglinear(alpha = 0, beta = 0.2)

}
