% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper.R
\name{expect_no_error}
\alias{expect_no_error}
\title{Helper functions}
\usage{
expect_no_error(expr)
}
\arguments{
\item{expr}{Expression.}
}
\description{
helper function that augments
\code{test_that::expect_no_error()} to expect no error.
Copied from the \code{R6} source code.
}
\details{
Small utility functions. Not to be exported to the user.
}
\keyword{internal}
