% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vectorized.R
\name{fix_vfunc_arguments}
\alias{fix_vfunc_arguments}
\title{helper function that fixes the arguments in the vectorised functions}
\usage{
fix_vfunc_arguments(
  lambda_mat = NULL,
  Lambda_mat = NULL,
  range_t = NULL,
  generate_lambda = TRUE,
  generate_Lambda = TRUE
)
}
\description{
helper function that fixes the arguments in the vectorised functions
}
