% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-linear.R
\name{Lambda_exp_form}
\alias{Lambda_exp_form}
\title{Definite integral of \code{l = exp(intercept + slope*t)} at time \code{t}
with \code{L(t0) = 0}}
\usage{
Lambda_exp_form(t, intercept, slope, t0)
}
\arguments{
\item{t}{(double) the time point}

\item{intercept}{(double) the intercept}

\item{slope}{(double) the slope}

\item{t0}{(double) the starting time}
}
\description{
Definite integral of \code{l = exp(intercept + slope*t)} starting at
\code{t0} -- only for \verb{l+}.
}
\keyword{internal}
