% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vectorized.R
\name{make_cumulative_Lambda_matrix}
\alias{make_cumulative_Lambda_matrix}
\title{Helper function for the vectorized versions of sampling functions.
Takes the usual ways that \code{lambda_mat} and \code{Lambda_mat} are specified
and returns \code{Lambda_mat}.}
\usage{
make_cumulative_Lambda_matrix(
  Lambda_mat = NULL,
  lambda_mat = NULL,
  interval_duration = NULL
)
}
\arguments{
\item{Lambda_mat}{a matrix of cumulative intensities or missing}

\item{lambda_mat}{a matrix of intensities or missing}

\item{interval_duration}{a vector with the same number of elements as the rows of \code{Lambda_mat}}
}
\value{
A matrix (r x 2), row-expanded if needed
}
\description{
Helper function for the vectorized versions of sampling functions.
Takes the usual ways that \code{lambda_mat} and \code{Lambda_mat} are specified
and returns \code{Lambda_mat}.
}
