% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateMatrixRegularStepClass.R
\name{make_cumulative_RateMatrixRegularStep}
\alias{make_cumulative_RateMatrixRegularStep}
\title{Make a RateMatrixRegularStep object cumulative}
\usage{
make_cumulative_RateMatrixRegularStep(x, interval_duration)
}
\arguments{
\item{x}{(RateMatrixRegularStep) RateMatrixRegularStep object}

\item{interval_duration}{(double, vector) the length of the intervals}
}
\value{
a RateMatrixRegularStep object (cumulative = TRUE)
}
\description{
Make a RateMatrixRegularStep object cumulative
}
