% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateMatrixRegularStepClass.R
\name{make_instantaneous_RateMatrixRegularStep}
\alias{make_instantaneous_RateMatrixRegularStep}
\title{Make a RateMatrixRegularStep object instantaneous}
\usage{
make_instantaneous_RateMatrixRegularStep(x, interval_duration)
}
\arguments{
\item{x}{(RateMatrixRegularStep) RateMatrixRegularStep object}

\item{interval_duration}{(double, vector) the length of the intervals}
}
\value{
a RateMatrixRegularStep object (cumulative = FALSE)
}
\description{
Make a RateMatrixRegularStep object instantaneous
}
