% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vectorized.R
\name{make_range_t_matrix}
\alias{make_range_t_matrix}
\title{Helper function for the vectorized versions of sampling functions.
Takes the usual ways that \code{range_t} is specified
(a 2-vector, a 1 x 2 or an r x 2 matrix) and
returns a r x 2 matrix.}
\usage{
make_range_t_matrix(range_t, n_rows)
}
\arguments{
\item{range_t}{a 2-vector, a 1 x 2 or an r x 2 matrix}

\item{n_rows}{the number of rows in the fully expanded matrix (\code{r})}
}
\value{
A matrix (r x 2), row-expanded if needed
}
\description{
Helper function for the vectorized versions of sampling functions.
Takes the usual ways that \code{range_t} is specified
(a 2-vector, a 1 x 2 or an r x 2 matrix) and
returns a r x 2 matrix.
}
\keyword{internal}
