% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vectorized.R
\name{mat_cumsum_columns}
\alias{mat_cumsum_columns}
\title{Return matrix with column-wise cumulative sum
No checks for arguments is done.}
\usage{
mat_cumsum_columns(X)
}
\arguments{
\item{X}{(matrix)}
}
\value{
matrix
}
\description{
Return matrix with column-wise cumulative sum
No checks for arguments is done.
}
\keyword{internal}
