% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_cumsum_columns.R
\name{matrix_cumsum_columns}
\alias{matrix_cumsum_columns}
\title{Usage: matrix_cumsum_columns( X )}
\usage{
matrix_cumsum_columns(X)
}
\description{
Usage: matrix_cumsum_columns( X )
}
\keyword{internal}
