% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_diff_columns_inplace.R
\name{matrix_diff_columns_inplace}
\alias{matrix_diff_columns_inplace}
\title{Usage: matrix_diff_columns_inplace( X )}
\usage{
matrix_diff_columns_inplace(X)
}
\description{
Usage: matrix_diff_columns_inplace( X )
}
\keyword{internal}
