% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalClass.R
\name{new_Interval}
\alias{new_Interval}
\title{Generate Interval (S3 class) object}
\usage{
new_Interval(range_t)
}
\arguments{
\item{x}{(vector, matrix) numeric values for the interval}
}
\value{
an Interval object
}
\description{
Generate Interval (S3 class) object
}
