% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateMatrixRegularStepClass.R
\name{new_RateMatrixRegularStep}
\alias{new_RateMatrixRegularStep}
\title{Create new RateMatrixRegularStep object}
\usage{
new_RateMatrixRegularStep(x, cumulative = logical())
}
\arguments{
\item{x}{(matrix) numeric values for the rate matrix}

\item{cumulative}{(logical) is the rate matrix cumulative?}
}
\value{
a RateMatrixRegularStep object
}
\description{
Create new RateMatrixRegularStep object
}
