% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhppp-package.R
\docType{package}
\name{nhppp-package}
\alias{nhppp}
\alias{nhppp-package}
\title{nhppp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Trikalinos \email{ttrikalin@mac.com} (\href{https://orcid.org/0000-0002-3990-1848}{ORCID})

}
\keyword{internal}
