% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_n_cumul_intensity.R
\name{nhppp_n_cumulative_intensity}
\alias{nhppp_n_cumulative_intensity}
\title{Simulate `size` samples from a non homogeneous Poisson Point Process (NHPPP) in the interval
   (t_min, t_max)}
\usage{
nhppp_n_cumulative_intensity(
  size,
  Lambda,
  Lambda_inv = NULL,
  range_t = c(0, 10),
  range_L = c(Lambda(range_t[1]), Lambda(range_t[2])),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{size}{int, the number of samples needed}

\item{Lambda}{(function, double vector) a continuous increasing R to R map
which is the integrated rate of the NHPPP}

\item{Lambda_inv}{(function, double vector) the inverse of `Lambda()`}

\item{range_t}{(vector, double) min and max of the time interval}

\item{range_L}{(vector, double) min and max of the transformed time interval}

\item{rng_stream}{(`rstream`) an `rstream`, `RNGClass` object or `NULL`.}
}
\value{
a vector of `size` event times
}
\description{
Sample NHPPP times using the order statistics method,
optionally using an `rstream` generator or a `Kystis` `RNGClass` object.
}
\examples{
x <- nhppp_n_cumulative_intensity(23, Lambda = function(t) t + cos(t) - 1)
}
