% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_n_intensity.R
\name{nhppp_n_intensity}
\alias{nhppp_n_intensity}
\title{Simulate `size` samples from a non homogeneous Poisson Point Process (NHPPP) from
   (t0, t_max) (thinning method)}
\usage{
nhppp_n_intensity(
  size,
  lambda,
  lambda_maj = NULL,
  range_t = c(0, 10),
  rng_stream = NULL
)
}
\arguments{
\item{size}{(double) the number of samples.}

\item{lambda}{(function) the instantaneous rate of the NHPPP.
A continuous function of time.}

\item{lambda_maj}{(double, vector) the intercept and optional slope of the majorizing
linear function in `range_t`.}

\item{range_t}{(vector, double) min and max of the time interval.}

\item{rng_stream}{(`rstream`) an `rstream` or `RNGClass` object, or `NULL`}
}
\value{
a vector of event times (t_); if no events realize,
        a vector of length 0
}
\description{
Sample NHPPP intensity times using the thinning method, optionally using
an `rstream` generator or a `Kystis` `RNGClass` object.
}
\examples{
x <- nhppp_n_intensity(23, lambda = function(t) 1 + sin(t))
}
