% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_n_intensity_linear.R
\name{nhppp_n_intensity_linear}
\alias{nhppp_n_intensity_linear}
\title{Simulate `size` samples from a non homogeneous Poisson Point Process (NHPPP) from
   (t_min, t_max) with linear intensity function (inversion method)}
\usage{
nhppp_n_intensity_linear(
  size,
  alpha = 1,
  beta = 0,
  range_t = c(0, 10),
  tol = 10^-6,
  rng_stream = NULL
)
}
\arguments{
\item{size}{(double) the number of samples}

\item{alpha}{(double) the intercept}

\item{beta}{(double) the slope}

\item{range_t}{(vector, double) min and max of the time interval}

\item{tol}{the probability that we will have more than
the drawn events in (t_min, t_max]}

\item{rng_stream}{(`rstream`) an `rstream` object.}
}
\value{
a vector of event times (t_); if no events realize,
        a vector of length 0
}
\description{
Sample NHPPP times from a linear intensity function
using the inversion method, optionally using an `rstream`
generator or a `Kystis` `RNGClass` object
}
\examples{
x <- nhppp_n_intensity_linear(alpha = 0, beta = 0.2)

}
