% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_t_cumul_intensity_orderstats.R
\name{nhppp_t_cumulative_intensity_orderstats}
\alias{nhppp_t_cumulative_intensity_orderstats}
\title{Simulate from a non homogeneous Poisson Point Process (NHPPP) from
   (t_min, t_max) (order statistics method)}
\usage{
nhppp_t_cumulative_intensity_orderstats(
  Lambda,
  Lambda_inv = NULL,
  range_t = c(0, 10),
  range_L = c(Lambda(range_t[1]), Lambda(range_t[2])),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{Lambda}{(function, double vector) a continuous increasing R to R map
which is the integrated rate of the NHPPP}

\item{Lambda_inv}{(function, double vector) the inverse of `Lambda()`}

\item{range_t}{(vector, double) min and max of the time interval}

\item{range_L}{(vector, double) min and max of the transformed time interval}

\item{rng_stream}{(`rstream`) an `rstream`, `RNGClass` object or `NULL`.}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times (t_); if no events realize,
        a vector of length 0
}
\description{
Sample NHPPP times using the order statistics method,
optionally using an `rstream` generator or a `Kystis` `RNGClass` object.
}
\examples{
x <- nhppp_t_cumulative_intensity_orderstats(Lambda = function(t) t + cos(t) - 1)
}
