% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_t_intensity.R
\name{nhppp_t_intensity_piecewise}
\alias{nhppp_t_intensity_piecewise}
\title{Simulate from a non homogeneous Poisson Point Process (NHPPP) from
   (t0, t_max) (thinning method) with piecewise constant_majorizer}
\usage{
nhppp_t_intensity_piecewise(
  lambda,
  lambda_maj_vector = lambda(1:10),
  times_vector = 0:10,
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{lambda}{(function) the instantaneous rate of the NHPPP.
A continuous function of time.}

\item{lambda_maj_vector}{(scalar, double) `K` constant majorizing rates, one per interval}

\item{times_vector}{(vector, double) `K+1` time points defining `K` intervals
of constant rates:
     `[t_1 = range_t[1],       t_2)`: the first interval
     `[t_k,                t_{k+1})`: the `k`-th interval
     `[t_{K}, t_{K+1} = range_t[2])`: the `K`-th (last) interval}

\item{rng_stream}{(`rstream`) an `rstream` or `RNGClass` object, or `NULL`}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times (t_); if no events realize,
        a vector of length 0
}
\description{
Sample NHPPP times using the thinning method, optionally using
an `rstream` generator or a `Kystis` `RNGClass` object.
}
\examples{
x <- nhppp_t_intensity_piecewise(lambda = function(t) exp(.02 * t))
}
