% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhppp_t_cumulative_intensity_piecewise.R
\name{nhppp_t_piecewise_regular_vec}
\alias{nhppp_t_piecewise_regular_vec}
\title{Vectorized sampling of piecewise constant cumulative intensities
with same interval lengths
Simulate a piecewise constant-rate Poisson Point Process over \verb{(t_min, t_max]} (inversion method)
where the intervals have the same length (are "regular").}
\usage{
nhppp_t_piecewise_regular_vec(
  Lambda_matrix,
  range_t = c(0, 10),
  rng_stream = NULL,
  tol = 10^-6,
  only1 = FALSE,
  zero_truncated = FALSE
)
}
\arguments{
\item{Lambda_matrix}{(scalar, double) \code{K} integrated rates, one per interval}

\item{range_t}{(vector, double) \code{t_min} and \code{t_max}}

\item{rng_stream}{an \code{rstream} or \code{RNGClass} object}

\item{only1}{boolean, draw at most 1 event time}

\item{zero_truncated}{boolean, draw at least 1 event time}
}
\value{
a vector of event times t
if no events realize, it will have 0 length
}
\description{
Vectorized sampling of piecewise constant cumulative intensities
with same interval lengths
Simulate a piecewise constant-rate Poisson Point Process over \verb{(t_min, t_max]} (inversion method)
where the intervals have the same length (are "regular").
}
\examples{
x <- nhppp_t_piecewise_regular_vec(Lambda_vector = 1:5)
}
