% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ppp.R
\name{ppp_t_piecewise}
\alias{ppp_t_piecewise}
\title{Simulate a piecewise constant-rate Poisson Point Process over (t_min, t_max] (inversion method)}
\usage{
ppp_t_piecewise(
  rates_vector = rep(1, 5),
  times_vector = c(0:5),
  rng_stream = NULL,
  only1 = FALSE,
  zero_truncated = FALSE
)
}
\arguments{
\item{rates_vector}{(scalar, double) \code{K} constant rates, one per interval}

\item{times_vector}{(vector, double) \code{K+1} time points defining \code{K} intervals
of constant rates:
\verb{[t_1 = range_t[1],       t_2)}: the first interval
\verb{[t_k,                t_\{k+1\})}: the \code{k}-th interval
\verb{[t_\{K\}, t_\{K+1\} = range_t[2])}: the \code{K}-th (last) interval}

\item{rng_stream}{an \code{rstream} or \code{RNGClass} object}

\item{only1}{boolean, draw at most 1 event time}

\item{zero_truncated}{boolean, draw at least 1 event time}
}
\value{
a vector of event times t
if no events realize, it will have 0 length
}
\description{
Simulate a piecewise constant-rate Poisson Point Process over (t_min, t_max] (inversion method)
}
\examples{
x <- ppp_t_piecewise(rates_vector = rep(1, 5), times_vector = c(0:5))
}
