% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helper.R
\name{read_code}
\alias{read_code}
\title{Read code from text file as string}
\usage{
read_code(codeFile)
}
\arguments{
\item{codeFile}{Path to file}
}
\value{
\code{codeFile} contents as a character string
}
\description{
Read code from text file as string
}
\keyword{internal}
