% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rngstream.R
\name{rng_stream_rexp}
\alias{rng_stream_rexp}
\title{Exponential random samples from \code{rstream} objects}
\usage{
rng_stream_rexp(size = 1, rate = 1, rng_stream = NULL)
}
\arguments{
\item{size}{Integer, number of samples}

\item{rate}{Positive number, the rate (i.e., 1/mean)}

\item{rng_stream}{(\code{rstream}) an \code{rstream} object or \code{NULL}}
}
\value{
a vector of exponential variates of size \code{size}
}
\description{
Sample from \code{rstream} objects
}
\examples{
rng_stream_rexp(10)
}
