% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rngstream.R
\name{rng_stream_rztpois}
\alias{rng_stream_rztpois}
\title{Zero-truncated Poisson random samples from \code{rstream} objects}
\usage{
rng_stream_rztpois(size = 1, lambda = 1, rng_stream = NULL)
}
\arguments{
\item{size}{Integer, number of samples}

\item{lambda}{Positive number, the mean of the original
(untruncated) Poisson distribution}

\item{rng_stream}{(\code{rstream}) an \code{rstream} object or \code{NULL}}
}
\value{
a vector of non zero counts of size \code{size}
}
\description{
Sample from \code{rstream} objects
}
\examples{
rng_stream_rztpois(10)
}
