% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ppp.R
\name{sim_ppp_n}
\alias{sim_ppp_n}
\title{Simulate n events from a homogeneous Poisson Point Process.}
\usage{
sim_ppp_n(n = 1, rate = 1, t_min = 0, rng_stream = NULL)
}
\arguments{
\item{n}{scalar number of samples}

\item{rate}{scalar instantaneous rate}

\item{t_min}{scalar for the starting time value}

\item{rng_stream}{an `rstream` or `RNGClass` object}
}
\value{
a vector with event times t (starting from t_min)
}
\description{
Simulate n events from a homogeneous Poisson Point Process.
}
\examples{
x <- sim_ppp_n(n = 10, rate = 1, t_min = 0)
}
