% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalClass.R
\name{validate_Interval}
\alias{validate_Interval}
\title{Validate an Interval object}
\usage{
validate_Interval(x)
}
\arguments{
\item{x}{(Interval) Interval object}
}
\value{
Interval object if valid
}
\description{
Validate an Interval object
}
