% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateMatrixRegularStepClass.R
\name{validate_RateMatrixRegularStep}
\alias{validate_RateMatrixRegularStep}
\title{Validate a RateMatrixRegularStep object}
\usage{
validate_RateMatrixRegularStep(x)
}
\arguments{
\item{x}{(RateMatrixRegularStep) RateMatrixRegularStep object}
}
\value{
RateMatrixRegularStep object if valid
}
\description{
Validate a RateMatrixRegularStep object
}
