% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdraw_intensity_step_regular_R.R
\name{vdraw_intensity_step_regular_R2}
\alias{vdraw_intensity_step_regular_R2}
\title{Vectorized sampling from a non homogeneous Poisson Point Process (NHPPP) from
an interval (thinning method) with piecewise constant_majorizers (R)}
\usage{
vdraw_intensity_step_regular_R2(
  lambda = NULL,
  lambda_args = NULL,
  lambda_majorizer = NULL,
  interval = NULL,
  tol = 10^-6,
  atmost1 = FALSE,
  force_zt_majorizer = FALSE,
  validate = FALSE,
  ...
)
}
\arguments{
\item{lambda}{(function) a vectorized intensity function, with one or two arguments.
The first is time. The optional second is a named list with additional arguments.}

\item{lambda_args}{(list) optional list of named arguments for \code{lambda()}}

\item{lambda_majorizer}{(RateMatrixRegularClass) instantaneous intensity object, one per interval, regular steps}

\item{interval}{(Interval) vectorized Interval object}

\item{tol}{(scalar, double) tolerance for the number of events}

\item{atmost1}{boolean, draw at most 1 event time}

\item{atleast1}{boolean, draw at least 1 event time in interval}
}
\value{
a matrix of event times (columns) per draw (rows)
NAs are structural empty spots
}
\description{
Vectorized sampling from a non homogeneous Poisson Point Process (NHPPP) from
an interval (thinning method) with piecewise constant_majorizers.
The majorizers are step functions over equal-length time intevals.
}
\examples{
Z <- vdraw_intensity_step_regular_R2(
  lambda = function(x, ...) 0.1 * x,
  lambda_majorizer = new_RateMatrixRegularStep(matrix(rep(1, 5), nrow = 1), FALSE)
)
}
