% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vppp_regularstep.R
\name{vppp_regularstep}
\alias{vppp_regularstep}
\title{Vectorized sampling of piecewise constant cumulative intensities
with same interval lengths}
\usage{
vppp_regularstep(Lambda_matrix, range_t = c(0, 10), tol = 10^-6, only1 = FALSE)
}
\arguments{
\item{Lambda_matrix}{(matrix) integrated rates, one per interval}

\item{range_t}{(vector, double) \code{t_min} and \code{t_max}}

\item{tol}{(scalar, double) tolerance for the number of events}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times t
if no events realize, it will have 0 length
}
\description{
Simulate a piecewise constant-rate Poisson Point Process over \verb{(t_min, t_max]} (inversion method)
where the intervals have the same length (are "regular").
}
\examples{
x <- vppp_regularstep(Lambda_matrix = matrix(1:5, nrow = 1))
}
