% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ztnhppp_t_intensity_exponential.R
\name{ztnhppp_t_intensity_exponential}
\alias{ztnhppp_t_intensity_exponential}
\title{Simulate from a zero-truncated non homogeneous Poisson Point Process (zt-NHPPP) from
   (t_min, t_max) with exponential intensity function (inversion method)}
\usage{
ztnhppp_t_intensity_exponential(
  alpha = 1,
  beta = 0,
  range_t = c(0, 10),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{alpha}{(double) the intercept in the exponent}

\item{beta}{(double) the slope in the exponent}

\item{range_t}{(vector, double) min and max of the time interval}

\item{rng_stream}{(`rstream`) an `rstream` object.}

\item{only1}{boolean, 1 event time}
}
\value{
a vector of at least 1 event times
}
\description{
Sample zt-NHPPP times from an exponential intensity function
using the inversion method, optionally using an `rstream`
generator or a `Kystis` `RNGClass` object
}
\examples{
x <- ztnhppp_t_intensity_exponential(alpha = 0, beta = 0.2)

}
