% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ztnhppp_t_intensity_linear.R
\name{ztnhppp_t_intensity_linear}
\alias{ztnhppp_t_intensity_linear}
\title{Simulate `size` samples from a zero-truncated non homogeneous Poisson Point Process (zt-NHPPP) from
   (t_min, t_max) with linear intensity function}
\usage{
ztnhppp_t_intensity_linear(
  alpha = 1,
  beta = 0,
  range_t = c(0, 10),
  rng_stream = NULL,
  only1 = FALSE
)
}
\arguments{
\item{alpha}{(double) the intercept}

\item{beta}{(double) the slope}

\item{range_t}{(vector, double) min and max of the time interval}

\item{rng_stream}{(`rstream`) an `rstream` object}

\item{only1}{(boolean) draw 1 event time}
}
\value{
a vector of event times (t_); if no events realize,
        a vector of length 0
}
\description{
Sample zero-truncated NHPPP times from a linear intensity function
using the inversion method, optionally using an `rstream`
generator or a `Kystis` `RNGClass` object
}
\examples{
x <- ztnhppp_t_intensity_linear(alpha = 0, beta = 0.2)

}
