% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztppp_t.R
\name{ztppp_t}
\alias{ztppp_t}
\title{Simulate a zero-truncated homogeneous Poisson Point Process over (t_min, t_max]}
\usage{
ztppp_t(range_t = c(0, 10), rate = 1, rng_stream = NULL, only1 = FALSE)
}
\arguments{
\item{range_t}{(vector, double) min and max of the time interval}

\item{rate}{(scalar, double) constant instantaneous rate}

\item{rng_stream}{an \code{rstream} object}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times of size \code{size}
}
\description{
Simulate a zero-truncated homogeneous Poisson Point Process over (t_min, t_max]
}
\examples{
x <- ztppp_t(range_t = c(0, 10), rate = 0.001)
}
