% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-linear.R
\name{Lambda_inv_exp_form}
\alias{Lambda_inv_exp_form}
\title{Inverse of the definite integral of \code{l = exp(intercept + slope*t)} at time \code{t}}
\usage{
Lambda_inv_exp_form(z, intercept, slope, t0)
}
\arguments{
\item{z}{(double) the value of integrated rate for which you want to find the time}

\item{intercept}{(double) the intercept}

\item{slope}{(double) the slope}

\item{t0}{(double) the starting time}
}
\description{
Inverse of the definite integral of \code{l = exp(intercept + slope*t)} only for \verb{l+}.
}
\keyword{internal}
