% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhppp-package.R
\docType{package}
\name{nhppp-package}
\alias{nhppp}
\alias{nhppp-package}
\title{nhppp: Simulating Nonhomogeneous Poisson Point Processes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulates events from one dimensional nonhomogeneous Poisson point processes (NHPPPs) as per Trikalinos and Sereda (2024, \doi{10.48550/arXiv.2402.00358}). Functions are based on three algorithms that provably sample from a target NHPPP: the time-transformation of a homogeneous Poisson process (of intensity one) via the inverse of the integrated intensity function (Cinlar E, "Theory of stochastic processes" (1975, ISBN:0486497996)); the generation of a Poisson number of order statistics from a fixed density function; and the thinning of a majorizing NHPPP via an acceptance-rejection scheme (Lewis PAW, Shedler, GS (1979) \doi{10.1002/nav.3800260304}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bladder-ca/nhppp}
  \item Report bugs at \url{https://github.com/bladder-ca/nhppp/issues}
}

}
\author{
\strong{Maintainer}: Thomas Trikalinos \email{thomas_trikalinos@brown.edu} (\href{https://orcid.org/0000-0002-3990-1848}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yuliia Sereda \email{sereda_yuliia@brown.edu} (\href{https://orcid.org/0000-0002-4017-4561}{ORCID})
}

}
\keyword{internal}
