% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_next_n.R
\name{ppp_next_n}
\alias{ppp_next_n}
\title{Simulate n events from a homogeneous Poisson Point Process.}
\usage{
ppp_next_n(n = 1, rate = 1, t_min = 0, rng_stream = deprecated())
}
\arguments{
\item{n}{scalar number of samples}

\item{rate}{scalar instantaneous rate}

\item{t_min}{scalar for the starting time value}

\item{rng_stream}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} an \code{rstream} object}
}
\value{
a vector with event times t (starting from t_min)
}
\description{
Simulate n events from a homogeneous Poisson Point Process.
}
\examples{
x <- ppp_next_n(n = 10, rate = 1, t_min = 0)
}
