% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_t_orderstat.R
\name{ppp_t_orderstat}
\alias{ppp_t_orderstat}
\title{Simulate a homogeneous Poisson Point Process over (t_min, t_max] (order statistics method)}
\usage{
ppp_t_orderstat(range_t = c(0, 10), rate = 1, rng_stream = NULL, only1 = FALSE)
}
\arguments{
\item{range_t}{(vector, double) min and max of the time interval}

\item{rate}{(scalar, double) constant instantaneous rate}

\item{rng_stream}{an \code{rstream} object}

\item{only1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times t
if no events realize, it will have 0 length
}
\description{
Simulate a homogeneous Poisson Point Process over (t_min, t_max] (order statistics method)
}
\examples{
x <- ppp_t_orderstat(range_t = c(0, 10), rate = 1)
}
