% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ppp.R
\name{ppp_t_piecewise}
\alias{ppp_t_piecewise}
\title{Simulate a piecewise constant-rate Poisson Point Process over (t_min, t_max] (inversion method)}
\usage{
ppp_t_piecewise(
  rates_vector = rep(1, 5),
  times_vector = c(0:5),
  rng_stream = NULL,
  only1 = FALSE,
  zero_truncated = FALSE
)
}
\arguments{
\item{rates_vector}{(scalar, double) `K` constant rates, one per interval}

\item{times_vector}{(vector, double) `K+1` time points defining `K` intervals
of constant rates:
     `[t_1 = range_t[1],       t_2)`: the first interval
     `[t_k,                t_{k+1})`: the `k`-th interval
     `[t_{K}, t_{K+1} = range_t[2])`: the `K`-th (last) interval}

\item{rng_stream}{an `rstream` or `RNGClass` object}

\item{only1}{boolean, draw at most 1 event time}

\item{zero_truncated}{boolean, draw at least 1 event time}
}
\value{
a vector of event times t
        if no events realize, it will have 0 length
}
\description{
Simulate a piecewise constant-rate Poisson Point Process over (t_min, t_max] (inversion method)
}
\examples{
x <- ppp_t_piecewise(rates_vector = rep(1, 5), times_vector = c(0:5))
}
