% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rngstream.R
\name{rng_stream_runif}
\alias{rng_stream_runif}
\title{Uniform random samples from \code{rstream} objects}
\usage{
rng_stream_runif(size = 1, minimum = 0, maximum = 1, rng_stream = NULL)
}
\arguments{
\item{size}{Integer, number of samples}

\item{minimum}{Lower bound}

\item{maximum}{Upper bound}

\item{rng_stream}{(\code{rstream}) an \code{rstream} object or \code{NULL}}
}
\value{
a vector of uniform variates of size \code{size}
}
\description{
Sample from \code{rstream} objects
}
\examples{
rng_stream_runif(10)
}
