% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rngstream.R
\name{rztpois}
\alias{rztpois}
\title{Zero-truncated Poisson random samples (basic R)}
\usage{
rztpois(size = 1, lambda = 1)
}
\arguments{
\item{size}{Integer, number of samples}

\item{lambda}{Positive number, the mean of the original
(untruncated) Poisson distribution}
}
\value{
a vector of non zero counts of size \code{size}
}
\description{
Sample zero-truncated Poisson random samples (basic R)
}
\examples{
rztpois(10, 1)
rztpois(10, 1:10)
}
