% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdraw_cumulative_intensity.R
\name{vdraw_cumulative_intensity}
\alias{vdraw_cumulative_intensity}
\title{Vectorized simulation from a non homogeneous Poisson Point Process (NHPPP) from
(t_min, t_max) (inversion method)}
\usage{
vdraw_cumulative_intensity(
  Lambda,
  Lambda_inv,
  Lambda_args = NULL,
  Lambda_inv_args = NULL,
  range_t = NULL,
  tol = 10^-6,
  atmost1 = FALSE
)
}
\arguments{
\item{Lambda}{(function, double vector) an increasing function
which is the integrated rate of the NHPPP. It shoudl take a vectorized argument t for times and an optional arguments list}

\item{Lambda_inv}{(function, double vector) the inverse of \code{Lambda()}, also in vectorized form}

\item{Lambda_args}{(list) optional arguments to pass to Lambda}

\item{Lambda_inv_args}{(list) optional arguments to pass to Lambda_inv()}

\item{range_t}{(vector/matrix double) min and max of the time interval. If a vecor of 2 elements, we assume that 1 series of points is drawn.
If a matrix, its rows are the number of point processes that should be drawn.}

\item{tol}{the tolerange for the calulations}

\item{atmost1}{boolean, draw at most 1 event time}
}
\value{
a vector of event times (t_); if no events realize,
a vector of length 0
}
\description{
Sample NHPPP times using the inversion method,
optionally using an \code{rstream} generator object
}
