% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdraw_intensity_step_regular_cpp.R
\name{vdraw_intensity_step_regular_cpp}
\alias{vdraw_intensity_step_regular_cpp}
\title{Vectorized sampling from a non homogeneous Poisson Point Process (NHPPP) given
the intensity function with piecewise constant_majorizers (C++)}
\usage{
vdraw_intensity_step_regular_cpp(
  lambda = NULL,
  lambda_args = NULL,
  Lambda_maj_matrix = NULL,
  lambda_maj_matrix = NULL,
  rate_matrix_t_min = NULL,
  rate_matrix_t_max = NULL,
  t_min = NULL,
  t_max = NULL,
  tol = 10^-6,
  atmost1 = FALSE,
  atmostB = NULL
)
}
\arguments{
\item{lambda}{(function) intensity function, vectorized}

\item{lambda_args}{(list) optional named list of arguments to pass to \code{lambda}.
If you have arguments for \code{lambda} that vary by draw, they should be passed as
a data.table named \code{vector_arguments}.}

\item{Lambda_maj_matrix}{(matrix) integrated intensity rates at the end of each interval}

\item{lambda_maj_matrix}{(matrix) intensity rates, one per interval}

\item{rate_matrix_t_min}{(scalar | vector | column matrix) is the lower bound
of the time interval for each row of (Lambda|lambda)_maj_matrix.
The length of this argument is the number of point processes that should be drawn.}

\item{rate_matrix_t_max}{(scalar | vector | column matrix) the upper bound
of the time interval for each row of (Lambda|lambda)_maj_matrix.
The length of this argument is the number of point processes that should be drawn.}

\item{t_min}{(scalar | vector | column matrix) is the lower bound
of a subinterval of (rate_matrix_t_min, rate_matrix_t_max]. If set,
times are sampled from the subinterval.
If omitted, it is equivalent to \code{rate_matrix_t_min}.}

\item{t_max}{(scalar | vector | column matrix) is the upper bound
of a subinterval of (rate_matrix_t_min, rate_matrix_t_max]. If set,
times are sampled from the subinterval.
If omitted, it is equivalent to \code{rate_matrix_t_max}.}

\item{tol}{(scalar, double) tolerance for the number of events}

\item{atmost1}{boolean, draw at most 1 event time}

\item{atmostB}{If not NULL, draw at most B (B>0) event times. NULL means ignore.}
}
\value{
a matrix of event times (columns) per draw (rows)
NAs are structural empty spots
}
\description{
Vectorized sampling from a non homogeneous Poisson Point Process (NHPPP) using the intensity
function and piecewise constant_majorizers.
The majorizers are step functions over equal-length time intevals.
}
\keyword{internal}
